+++ example/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @react-native-web-adapter/example
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [455be68]
+  - @preflower/react-native-web-maps@0.2.0
+
+## 0.0.6
+
+### Patch Changes
+
+- 8448e56: update toast style that add horizontal spacing
+- Updated dependencies [8448e56]
+  - react-native-web-simple-toast@0.0.3
+
+## 0.0.5
+
+### Patch Changes
+
+- f668610: add Polyline and Polygon support
+- Updated dependencies [f668610]
+  - @preflower/react-native-web-maps@0.1.0
+
+## 0.0.4
+
+### Patch Changes
+
+- 1ba8ab9: chore: update README
+- Updated dependencies [2641516]
+- Updated dependencies [1ba8ab9]
+  - @preflower/react-native-web-maps@0.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- 16fd6f6: fix MapView not default output error
+- Updated dependencies [16fd6f6]
+  - @preflower/react-native-web-maps@0.0.2
+
+## 0.0.2
+
+### Patch Changes
+
+- dae96b5: Publish @preflower/react-native-web-maps
+- Updated dependencies [dae96b5]
+  - @preflower/react-native-web-maps@0.0.1
+
 ## 0.0.1
 
 ### Patch Changes