+++ packages/react-native-web-simple-toast/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-native-web-simple-toast
 
+## 0.0.3
+
+### Patch Changes
+
+- 8448e56: update toast style that add horizontal spacing
+
+## 0.0.2
+
+### Patch Changes
+
+- 14c3188: output develop information to the production env incorrectly
+
 ## 0.0.1
 
 ### Patch Changes